api/requirements.txt
setuptools>=65.5.1
wheel>=0.38.0
fastapi>=0.104.1
uvicorn>=0.23.2
python-multipart>=0.0.6
deepface>=0.0.79
Pillow==10.0.0
pydantic==2.4.2
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
opencv-python-headless==4.8.0.76
tensorflow-cpu==2.13.0
keras>=2.12.0
h5py>=3.7.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn==1.3.0
openpyxl>=3.0.0
flask==2.3.3
gunicorn==21.2.0
mtcnn==0.1.1 